scripts/generate-load-requirements.txt
python-dotenv
pandas
psycopg2-binary
sqlalchemy==1.4.46
